+++ src/main/java/HttpRequest.java
@@ -1,0 +1,4 @@
+import java.util.Map;
+
+record HttpRequest(String method, String path, Map<String, String> headers, String body) {
+}
